@@ -12,14 +12,14 @@
     "@tuleap/plugin-agiledashboard-burnup-data-transformer": "workspace:*",
     "@tuleap/tlp-fetch": "workspace:*",
     "@tuleap/tlp-swatch-colors": "workspace:*",
-    "@tuleap/vue2-gettext-composition-helper": "workspace:^",
-    "@tuleap/vue2-gettext-init": "workspace:*",
+    "@tuleap/vue3-gettext-init": "workspace:*",
     "@vue/composition-api": "^1.7.2",
     "d3-array": "^3.0.1",
     "d3-selection": "^3.0.0",
     "d3-shape": "^3.0.1",
     "pinia": "^2.0.21",
-    "vue": "^2.7.8",
+    "vue": "^3.3.4",
+    "vue3-gettext": "^2.4.0",
     "vue-dompurify-html": "^4.0.0"
   },
   "devDependencies": {
@@ -30,8 +30,8 @@
     "@types/d3-selection": "^2.0.0",
     "@types/d3-shape": "^2.1.0",
     "@types/sprintf-js": "^1.1.2",
-    "@vitejs/plugin-vue2": "^2.0.0",
-    "@vue/test-utils": "^1.3.6",
+    "@vitejs/plugin-vue": "^4.2.3",
+    "@vue/test-utils": "^2.2.10",
     "vue-class-component": "^7.2.6",
     "vue-property-decorator": "^9.1.2"
   },
