@@ -45,7 +45,7 @@
 import { computed } from "vue";
 import ReleaseDisplayer from "../WhatsHotSection/ReleaseDisplayer.vue";
 import { useStore } from "../../stores/root";
-import { useGettext } from "@tuleap/vue2-gettext-composition-helper";
+import { useGettext } from "vue3-gettext";
 
 const root_store = useStore();
 
