@@ -17,16 +17,16 @@
  * along with Tuleap. If not, see <http://www.gnu.org/licenses/>.
  */
 
-import type { Wrapper } from "@vue/test-utils";
+import type { VueWrapper } from "@vue/test-utils";
 import { shallowMount } from "@vue/test-utils";
 import RoadmapSection from "./RoadmapSection.vue";
-import { createReleaseWidgetLocalVue } from "../../helpers/local-vue-for-test";
 import { createTestingPinia } from "@pinia/testing";
 import { defineStore } from "pinia";
+import { getGlobalTestOptions } from "../../helpers/global-options-for-test";
 
 const project_id = 102;
 
-async function getPersonalWidgetInstance(): Promise<Wrapper<Vue, Element>> {
+function getPersonalWidgetInstance(): VueWrapper<InstanceType<typeof RoadmapSection>> {
     const useStore = defineStore("root", {
         state: () => ({
             is_loading: false,
@@ -40,19 +40,20 @@ async function getPersonalWidgetInstance(): Promise<Wrapper<Vue, Element>> {
     useStore(pinia);
 
     const component_options = {
-        localVue: await createReleaseWidgetLocalVue(),
+        global: {
+            ...getGlobalTestOptions(pinia),
+        },
         propsData: {
             label_tracker_planning: "sprint",
         },
-        pinia,
     };
 
     return shallowMount(RoadmapSection, component_options);
 }
 
 describe("RoadmapSection", () => {
-    it("Given user display widget, Then a good link to top planning of the project is rendered", async () => {
-        const wrapper = await getPersonalWidgetInstance();
+    it("Given user display widget, Then a good link to top planning of the project is rendered", () => {
+        const wrapper = getPersonalWidgetInstance();
 
         expect(wrapper.element).toMatchSnapshot();
     });
