@@ -17,22 +17,22 @@
  * along with Tuleap. If not, see <http://www.gnu.org/licenses/>.
  */
 
-import type { Wrapper } from "@vue/test-utils";
+import type { VueWrapper } from "@vue/test-utils";
 import { shallowMount } from "@vue/test-utils";
 import ReleaseBadgesAllSprints from "./ReleaseBadgesAllSprints.vue";
 import type { MilestoneData, TrackerProjectLabel } from "../../../type";
-import { createReleaseWidgetLocalVue } from "../../../helpers/local-vue-for-test";
 import { createTestingPinia } from "@pinia/testing";
 import { defineStore } from "pinia";
+import { getGlobalTestOptions } from "../../../helpers/global-options-for-test";
 
 const total_sprint = 10;
 const initial_effort = 10;
 
 describe("ReleaseBadgesAllSprints", () => {
-    async function getPersonalWidgetInstance(
+    function getPersonalWidgetInstance(
         user_can_view_sub_milestones_planning: boolean,
         trackers: Array<TrackerProjectLabel>,
-    ): Promise<Wrapper<Vue, Element>> {
+    ): VueWrapper<InstanceType<typeof ReleaseBadgesAllSprints>> {
         const release_data = {
             id: 2,
             total_sprint,
@@ -55,32 +55,33 @@ describe("ReleaseBadgesAllSprints", () => {
         useStore(pinia);
 
         return shallowMount(ReleaseBadgesAllSprints, {
-            localVue: await createReleaseWidgetLocalVue(),
+            global: {
+                ...getGlobalTestOptions(pinia),
+            },
             propsData: { release_data, isPastRelease: false },
-            pinia,
         });
     }
 
     describe("Display number of sprint", () => {
-        it("When there is a tracker, Then number of sprint is displayed", async () => {
+        it("When there is a tracker, Then number of sprint is displayed", () => {
             const trackers = [
                 {
                     label: "Sprint1",
                 },
             ];
-            const wrapper = await getPersonalWidgetInstance(true, trackers);
+            const wrapper = getPersonalWidgetInstance(true, trackers);
 
             expect(wrapper.get("[data-test=badge-sprint]").text()).toBe("10 Sprint1");
         });
 
-        it("When there isn't tracker, Then there is no link", async () => {
-            const wrapper = await getPersonalWidgetInstance(true, []);
+        it("When there isn't tracker, Then there is no link", () => {
+            const wrapper = getPersonalWidgetInstance(true, []);
 
             expect(wrapper.find("[data-test=badge-sprint]").exists()).toBe(false);
         });
 
-        it("When the user can't see the tracker, Then number of sprint is not displayed", async () => {
-            const wrapper = await getPersonalWidgetInstance(false, []);
+        it("When the user can't see the tracker, Then number of sprint is not displayed", () => {
+            const wrapper = getPersonalWidgetInstance(false, []);
 
             expect(wrapper.find("[data-test=badge-sprint]").exists()).toBe(false);
         });
