@@ -17,21 +17,21 @@
  * along with Tuleap. If not, see <http://www.gnu.org/licenses/>.
  */
 
-import type { Wrapper } from "@vue/test-utils";
+import type { VueWrapper } from "@vue/test-utils";
 import { shallowMount } from "@vue/test-utils";
 import ReleaseBadgesClosedSprints from "./ReleaseBadgesClosedSprints.vue";
 import type { MilestoneData, TrackerProjectLabel } from "../../../type";
-import { createReleaseWidgetLocalVue } from "../../../helpers/local-vue-for-test";
 import { createTestingPinia } from "@pinia/testing";
 import { defineStore } from "pinia";
+import { getGlobalTestOptions } from "../../../helpers/global-options-for-test";
 
 const total_sprint = 10;
 
 describe("ReleaseBadgesClosedSprints", () => {
-    async function getPersonalWidgetInstance(
+    function getPersonalWidgetInstance(
         user_can_view_sub_milestones_planning: boolean,
         release_data: MilestoneData,
-    ): Promise<Wrapper<Vue, Element>> {
+    ): VueWrapper<InstanceType<typeof ReleaseBadgesClosedSprints>> {
         const useStore = defineStore("root", {
             state: () => ({
                 user_can_view_sub_milestones_planning,
@@ -41,14 +41,15 @@ describe("ReleaseBadgesClosedSprints", () => {
         useStore(pinia);
 
         return shallowMount(ReleaseBadgesClosedSprints, {
-            localVue: await createReleaseWidgetLocalVue(),
+            global: {
+                ...getGlobalTestOptions(pinia),
+            },
             propsData: { release_data },
-            pinia,
         });
     }
 
     describe("Display total of closed sprints", () => {
-        it("When there are some closed sprints, Then the total is displayed", async () => {
+        it("When there are some closed sprints, Then the total is displayed", () => {
             const release_data = {
                 id: 2,
                 total_sprint,
@@ -66,12 +67,12 @@ describe("ReleaseBadgesClosedSprints", () => {
                 },
             } as MilestoneData;
 
-            const wrapper = await getPersonalWidgetInstance(true, release_data);
+            const wrapper = getPersonalWidgetInstance(true, release_data);
 
             expect(wrapper.find("[data-test=total-closed-sprints]").exists()).toBe(true);
         });
 
-        it("When the total of closed sprints is null, Then the total is not displayed", async () => {
+        it("When the total of closed sprints is null, Then the total is not displayed", () => {
             const release_data = {
                 id: 2,
                 total_sprint,
@@ -89,12 +90,12 @@ describe("ReleaseBadgesClosedSprints", () => {
                 },
             } as MilestoneData;
 
-            const wrapper = await getPersonalWidgetInstance(true, release_data);
+            const wrapper = getPersonalWidgetInstance(true, release_data);
 
             expect(wrapper.find("[data-test=total-closed-sprints]").exists()).toBe(false);
         });
 
-        it("When the total of closed sprints is 0, Then the total is displayed", async () => {
+        it("When the total of closed sprints is 0, Then the total is displayed", () => {
             const release_data = {
                 id: 2,
                 total_sprint,
@@ -112,12 +113,12 @@ describe("ReleaseBadgesClosedSprints", () => {
                 },
             } as MilestoneData;
 
-            const wrapper = await getPersonalWidgetInstance(true, release_data);
+            const wrapper = getPersonalWidgetInstance(true, release_data);
 
             expect(wrapper.find("[data-test=total-closed-sprints]").exists()).toBe(true);
         });
 
-        it("When there is no trackers of sprints, Then the total is not displayed", async () => {
+        it("When there is no trackers of sprints, Then the total is not displayed", () => {
             const release_data = {
                 id: 2,
                 total_sprint,
@@ -131,12 +132,12 @@ describe("ReleaseBadgesClosedSprints", () => {
                 },
             } as MilestoneData;
 
-            const wrapper = await getPersonalWidgetInstance(true, release_data);
+            const wrapper = getPersonalWidgetInstance(true, release_data);
 
             expect(wrapper.find("[data-test=total-closed-sprints]").exists()).toBe(false);
         });
 
-        it("When the user can't see the tracker's label, Then the total is not displayed", async () => {
+        it("When the user can't see the tracker's label, Then the total is not displayed", () => {
             const release_data = {
                 id: 2,
                 total_sprint,
@@ -154,7 +155,7 @@ describe("ReleaseBadgesClosedSprints", () => {
                 },
             } as MilestoneData;
 
-            const wrapper = await getPersonalWidgetInstance(false, release_data);
+            const wrapper = getPersonalWidgetInstance(false, release_data);
 
             expect(wrapper.find("[data-test=total-closed-sprints]").exists()).toBe(false);
         });
