@@ -17,20 +17,20 @@
  * along with Tuleap. If not, see <http://www.gnu.org/licenses/>.
  */
 
-import type { Wrapper } from "@vue/test-utils";
+import type { VueWrapper } from "@vue/test-utils";
 import { shallowMount } from "@vue/test-utils";
 import ReleaseBadgesDisplayerIfOnlyClosedSprints from "./ReleaseBadgesDisplayerIfOnlyClosedSprints.vue";
 import type { MilestoneData } from "../../../type";
-import { createReleaseWidgetLocalVue } from "../../../helpers/local-vue-for-test";
 import ReleaseOthersBadges from "./ReleaseOthersBadges.vue";
 import ReleaseBadgesClosedSprints from "./ReleaseBadgesClosedSprints.vue";
 import { createTestingPinia } from "@pinia/testing";
 import { defineStore } from "pinia";
+import { getGlobalTestOptions } from "../../../helpers/global-options-for-test";
 
 describe("ReleaseBadgesDisplayerIfOnlyClosedSprints", () => {
-    async function getPersonalWidgetInstance(
+    function getPersonalWidgetInstance(
         user_can_view_sub_milestones_planning: boolean,
-    ): Promise<Wrapper<Vue, Element>> {
+    ): VueWrapper<InstanceType<typeof ReleaseBadgesDisplayerIfOnlyClosedSprints>> {
         const useStore = defineStore("root", {
             state: () => ({
                 user_can_view_sub_milestones_planning,
@@ -40,25 +40,26 @@ describe("ReleaseBadgesDisplayerIfOnlyClosedSprints", () => {
         useStore(pinia);
 
         return shallowMount(ReleaseBadgesDisplayerIfOnlyClosedSprints, {
-            localVue: await createReleaseWidgetLocalVue(),
+            global: {
+                ...getGlobalTestOptions(pinia),
+            },
             propsData: {
                 release_data: {
                     id: 2,
                 } as MilestoneData,
             },
-            pinia,
         });
     }
 
     describe("Display closed sprints and others badges", () => {
-        it("When the component is rendered, Then ReleaseBadgesClosedSprints and ReleaseOthersBadges are rendered", async () => {
-            const wrapper = await getPersonalWidgetInstance(true);
+        it("When the component is rendered, Then ReleaseBadgesClosedSprints and ReleaseOthersBadges are rendered", () => {
+            const wrapper = getPersonalWidgetInstance(true);
 
             expect(wrapper.findComponent(ReleaseBadgesClosedSprints).exists()).toBe(true);
             expect(wrapper.findComponent(ReleaseOthersBadges).exists()).toBe(true);
         });
-        it("If the user can't see sprints' tracker, Then ReleaseBadgesClosedSprints is not rendered", async () => {
-            const wrapper = await getPersonalWidgetInstance(false);
+        it("If the user can't see sprints' tracker, Then ReleaseBadgesClosedSprints is not rendered", () => {
+            const wrapper = getPersonalWidgetInstance(false);
 
             expect(wrapper.findComponent(ReleaseBadgesClosedSprints).exists()).toBe(false);
             expect(wrapper.findComponent(ReleaseOthersBadges).exists()).toBe(true);
