@@ -17,20 +17,20 @@
  * along with Tuleap. If not, see <http://www.gnu.org/licenses/>.
  */
 
-import type { Wrapper } from "@vue/test-utils";
+import type { VueWrapper } from "@vue/test-utils";
 import { shallowMount } from "@vue/test-utils";
 import ReleaseBadgesOpenSprint from "./ReleaseBadgesOpenSprint.vue";
 import type { MilestoneData } from "../../../type";
-import { createReleaseWidgetLocalVue } from "../../../helpers/local-vue-for-test";
 import ReleaseButtonsDescription from "../ReleaseDescription/ReleaseButtonsDescription.vue";
+import { getGlobalTestOptions } from "../../../helpers/global-options-for-test";
 
 const total_sprint = 10;
 
 describe("ReleaseBadgesOpenSprint", () => {
-    async function getPersonalWidgetInstance(
+    function getPersonalWidgetInstance(
         start_date: string | null,
         end_date: string | null,
-    ): Promise<Wrapper<Vue, Element>> {
+    ): VueWrapper<InstanceType<typeof ReleaseBadgesOpenSprint>> {
         const sprint_data = {
             label: "mile",
             id: 2,
@@ -51,7 +51,9 @@ describe("ReleaseBadgesOpenSprint", () => {
         } as MilestoneData;
 
         return shallowMount(ReleaseBadgesOpenSprint, {
-            localVue: await createReleaseWidgetLocalVue(),
+            global: {
+                ...getGlobalTestOptions(),
+            },
             propsData: {
                 sprint_data,
                 isPastRelease: false,
@@ -60,67 +62,61 @@ describe("ReleaseBadgesOpenSprint", () => {
     }
 
     describe("Display sprint data", () => {
-        it("When the component is rendered, Then the label of sprint is displayed", async () => {
-            const wrapper = await getPersonalWidgetInstance(null, null);
+        it("When the component is rendered, Then the label of sprint is displayed", () => {
+            const wrapper = getPersonalWidgetInstance(null, null);
 
             expect(wrapper.get("[data-test=sprint-label]").text()).toBe("mile");
         });
 
-        it("When the component is rendered, Then ReleaseBadgesButtonOpenSprint is rendered", async () => {
-            const wrapper = await getPersonalWidgetInstance(null, null);
+        it("When the component is rendered, Then ReleaseBadgesButtonOpenSprint is rendered", () => {
+            const wrapper = getPersonalWidgetInstance(null, null);
 
             expect(wrapper.findComponent(ReleaseButtonsDescription).exists()).toBe(true);
         });
 
-        it("When a release is not in progress, Then the badge is outline", async () => {
-            const wrapper = await getPersonalWidgetInstance(null, null);
+        it("When a release is not in progress, Then the badge is outline", () => {
+            const wrapper = getPersonalWidgetInstance(null, null);
 
             expect(wrapper.get("[data-test=sprint-label]").attributes("class")).toContain(
                 "tlp-badge-outline",
             );
         });
 
-        it("When a release is has no end date set, Then the badge is outline", async () => {
-            const wrapper = await getPersonalWidgetInstance(null, new Date().toString());
+        it("When a release is has no end date set, Then the badge is outline", () => {
+            const wrapper = getPersonalWidgetInstance(null, new Date().toString());
 
             expect(wrapper.get("[data-test=sprint-label]").attributes("class")).toContain(
                 "tlp-badge-outline",
             );
         });
 
-        it("When a release is has no start date set, Then the badge is outline", async () => {
-            const wrapper = await getPersonalWidgetInstance(new Date().toString(), null);
+        it("When a release is has no start date set, Then the badge is outline", () => {
+            const wrapper = getPersonalWidgetInstance(new Date().toString(), null);
 
             expect(wrapper.get("[data-test=sprint-label]").attributes("class")).toContain(
                 "tlp-badge-outline",
             );
         });
 
-        it("When a release is in progress, Then the badge is not outline", async () => {
+        it("When a release is in progress, Then the badge is not outline", () => {
             const start_date = new Date();
             const end_date = new Date();
             end_date.setDate(end_date.getDate() + 1);
 
-            const wrapper = await getPersonalWidgetInstance(
-                start_date.toString(),
-                end_date.toString(),
-            );
+            const wrapper = getPersonalWidgetInstance(start_date.toString(), end_date.toString());
 
             expect(wrapper.get("[data-test=sprint-label]").attributes("class")).not.toContain(
                 "tlp-badge-outline",
             );
         });
 
-        it("When a release is in past, Then the badge is not outline", async () => {
+        it("When a release is in past, Then the badge is not outline", () => {
             const start_date = new Date();
             start_date.setDate(1234567890);
             const end_date = new Date();
             end_date.setDate(end_date.getDate() + 1);
 
-            const wrapper = await getPersonalWidgetInstance(
-                start_date.toString(),
-                end_date.toString(),
-            );
+            const wrapper = getPersonalWidgetInstance(start_date.toString(), end_date.toString());
 
             expect(wrapper.get("[data-test=sprint-label]").attributes("class")).toContain(
                 "tlp-badge-outline",
