@@ -17,13 +17,13 @@
  * along with Tuleap. If not, see <http://www.gnu.org/licenses/>.
  */
 
-import type { Wrapper } from "@vue/test-utils";
-import { shallowMount } from "@vue/test-utils";
+import type { VueWrapper } from "@vue/test-utils";
+import { mount } from "@vue/test-utils";
 import ReleaseOthersBadges from "./ReleaseOthersBadges.vue";
 import type { MilestoneData, Pane, TrackerProjectLabel } from "../../../type";
-import { createReleaseWidgetLocalVue } from "../../../helpers/local-vue-for-test";
 import { createTestingPinia } from "@pinia/testing";
 import { defineStore } from "pinia";
+import { getGlobalTestOptions } from "../../../helpers/global-options-for-test";
 
 let release_data: MilestoneData & Required<Pick<MilestoneData, "planning">>;
 const total_sprint: number | null = 10;
@@ -60,9 +60,9 @@ let resources = {
 };
 
 describe("ReleaseOthersBadges", () => {
-    async function getPersonalWidgetInstance(
+    function getPersonalWidgetInstance(
         user_can_view_sub_milestones_planning = false,
-    ): Promise<Wrapper<Vue, Element>> {
+    ): VueWrapper<InstanceType<typeof ReleaseOthersBadges>> {
         release_data = {
             id: 2,
             capacity,
@@ -84,27 +84,28 @@ describe("ReleaseOthersBadges", () => {
         useStore(pinia);
 
         const component_options = {
-            localVue: await createReleaseWidgetLocalVue(),
-            pinia,
+            global: {
+                ...getGlobalTestOptions(pinia),
+            },
             propsData: {
                 release_data,
             },
         };
 
-        return shallowMount(ReleaseOthersBadges, component_options);
+        return mount(ReleaseOthersBadges, component_options);
     }
 
     describe("Display points of initial effort", () => {
-        it("When there is an initial effort, Then the points of initial effort are displayed", async () => {
-            const wrapper = await getPersonalWidgetInstance();
+        it("When there is an initial effort, Then the points of initial effort are displayed", () => {
+            const wrapper = getPersonalWidgetInstance();
 
             expect(wrapper.find("[data-test=initial-effort-not-empty]").exists()).toBe(true);
             expect(wrapper.find("[data-test=initial-effort-empty]").exists()).toBe(false);
         });
 
-        it("When there is initial effort but null, Then the points of initial effort are 'N/A'", async () => {
+        it("When there is initial effort but null, Then the points of initial effort are 'N/A'", () => {
             initial_effort = null;
-            const wrapper = await getPersonalWidgetInstance();
+            const wrapper = getPersonalWidgetInstance();
 
             expect(wrapper.find("[data-test=initial-effort-not-empty]").exists()).toBe(false);
             expect(wrapper.find("[data-test=initial-effort-empty]").exists()).toBe(true);
@@ -112,37 +113,37 @@ describe("ReleaseOthersBadges", () => {
     });
 
     describe("Display points of capacity", () => {
-        it("When there are points of capacity, Then the points of capacity are displayed", async () => {
-            const wrapper = await getPersonalWidgetInstance();
+        it("When there are points of capacity, Then the points of capacity are displayed", () => {
+            const wrapper = getPersonalWidgetInstance();
 
             expect(wrapper.find("[data-test=capacity-not-empty]").exists()).toBe(true);
             expect(wrapper.find("[data-test=capacity-empty]").exists()).toBe(false);
         });
 
-        it("When there are points of capacity but null, Then the points of capacity are 'N/A'", async () => {
+        it("When there are points of capacity but null, Then the points of capacity are 'N/A'", () => {
             capacity = null;
-            const wrapper = await getPersonalWidgetInstance();
+            const wrapper = getPersonalWidgetInstance();
 
             expect(wrapper.find("[data-test=capacity-not-empty]").exists()).toBe(false);
             expect(wrapper.find("[data-test=capacity-empty]").exists()).toBe(true);
         });
 
-        it("Given initial effort is bigger than capacity, Then the initial effort badge has a warning style", async () => {
+        it("Given initial effort is bigger than capacity, Then the initial effort badge has a warning style", () => {
             initial_effort = 100;
             capacity = 50;
 
-            const wrapper = await getPersonalWidgetInstance();
+            const wrapper = getPersonalWidgetInstance();
 
             expect(wrapper.get("[data-test=initial_effort_badge]").classes()).toContain(
                 "tlp-badge-warning",
             );
         });
 
-        it("Given initial effort is smaller than capacity, Then the initial effort badge has primary style", async () => {
+        it("Given initial effort is smaller than capacity, Then the initial effort badge has primary style", () => {
             capacity = 100;
             initial_effort = 50;
 
-            const wrapper = await getPersonalWidgetInstance();
+            const wrapper = getPersonalWidgetInstance();
 
             expect(wrapper.get("[data-test=initial_effort_badge]").classes()).toContain(
                 "tlp-badge-primary",
@@ -150,8 +151,8 @@ describe("ReleaseOthersBadges", () => {
         });
     });
 
-    it("Given user display widget, Then a good link to sprint planning is renderer", async () => {
-        const wrapper = await getPersonalWidgetInstance(true);
+    it("Given user display widget, Then a good link to sprint planning is renderer", () => {
+        const wrapper = getPersonalWidgetInstance(true);
         expect(wrapper.get("[data-test=planning-link]").attributes("href")).toEqual(
             "/plugins/agiledashboard/?group_id=" +
                 encodeURIComponent(project_id) +
@@ -163,12 +164,12 @@ describe("ReleaseOthersBadges", () => {
         );
     });
 
-    it("When the user can't see the subplanning, Then he can't see the planning link", async () => {
-        const wrapper = await getPersonalWidgetInstance(false);
+    it("When the user can't see the subplanning, Then he can't see the planning link", () => {
+        const wrapper = getPersonalWidgetInstance(false);
         expect(wrapper.find("[data-test=planning-link]").exists()).toBe(false);
     });
 
-    it("When there isn't sub-planning, Then there isn't any link to sub-planning", async () => {
+    it("When there isn't sub-planning, Then there isn't any link to sub-planning", () => {
         resources = {
             milestones: {
                 accept: {
@@ -181,7 +182,7 @@ describe("ReleaseOthersBadges", () => {
             },
         };
 
-        const wrapper = await getPersonalWidgetInstance();
+        const wrapper = getPersonalWidgetInstance();
         expect(wrapper.find("[data-test=planning-link]").exists()).toBe(false);
     });
 });
