@@ -17,10 +17,10 @@
  * along with Tuleap. If not, see <http://www.gnu.org/licenses/>.
  */
 
-import type { Wrapper } from "@vue/test-utils";
+import type { VueWrapper } from "@vue/test-utils";
 import { shallowMount } from "@vue/test-utils";
-import { createReleaseWidgetLocalVue } from "../../../../helpers/local-vue-for-test";
 import ChartError from "./ChartError.vue";
+import { getGlobalTestOptions } from "../../../../helpers/global-options-for-test";
 
 const message_error_duration = "'duration' field is empty or invalid.";
 const message_error_start_date = "'start_date' field is empty or invalid.";
@@ -28,11 +28,11 @@ const message_error_under_calculation =
     "Burndown is under calculation. It will be available in a few minutes.";
 
 describe("ChartError", () => {
-    async function getPersonalWidgetInstance(
+    function getPersonalWidgetInstance(
         has_error_duration: boolean,
         has_error_start_date: boolean,
         is_under_calculation: boolean,
-    ): Promise<Wrapper<Vue, Element>> {
+    ): VueWrapper<InstanceType<typeof ChartError>> {
         return shallowMount(ChartError, {
             propsData: {
                 has_error_duration,
@@ -42,7 +42,9 @@ describe("ChartError", () => {
                 message_error_start_date,
                 message_error_under_calculation,
             },
-            localVue: await createReleaseWidgetLocalVue(),
+            global: {
+                ...getGlobalTestOptions(),
+            },
         });
     }
 
@@ -54,7 +56,7 @@ describe("ChartError", () => {
         [false, false, false, false, false, false],
     ])(
         `Error message %s`,
-        async (
+        (
             has_error_duration: boolean,
             has_error_start_date: boolean,
             is_under_calculation: boolean,
@@ -62,7 +64,7 @@ describe("ChartError", () => {
             display_start_date_error: boolean,
             display_calculation_error: boolean,
         ) => {
-            const wrapper = await getPersonalWidgetInstance(
+            const wrapper = getPersonalWidgetInstance(
                 has_error_duration,
                 has_error_start_date,
                 is_under_calculation,
