@@ -17,19 +17,19 @@
  * along with Tuleap. If not, see <http://www.gnu.org/licenses/>.
  */
 
-import type { Wrapper } from "@vue/test-utils";
+import type { VueWrapper } from "@vue/test-utils";
 import { shallowMount } from "@vue/test-utils";
 import type { MilestoneData } from "../../../../type";
-import { createReleaseWidgetLocalVue } from "../../../../helpers/local-vue-for-test";
 import TestManagement from "./TestManagement.vue";
 import { createTestingPinia } from "@pinia/testing";
 import { defineStore } from "pinia";
+import { getGlobalTestOptions } from "../../../../helpers/global-options-for-test";
 
 let release_data: MilestoneData;
 const project_id = 100;
 
 describe("TestManagement", () => {
-    async function getPersonalWidgetInstance(): Promise<Wrapper<Vue, Element>> {
+    function getPersonalWidgetInstance(): VueWrapper<InstanceType<typeof TestManagement>> {
         const useStore = defineStore("root", {
             state: () => ({
                 label_tracker_planning: "Releases",
@@ -40,12 +40,13 @@ describe("TestManagement", () => {
         useStore(pinia);
 
         return shallowMount(TestManagement, {
-            localVue: await createReleaseWidgetLocalVue(),
+            global: {
+                ...getGlobalTestOptions(pinia),
+            },
             propsData: {
                 release_data,
                 campaign: release_data.campaign,
             },
-            pinia,
         });
     }
 
@@ -74,7 +75,7 @@ describe("TestManagement", () => {
         } as MilestoneData;
     });
 
-    it("When there is not campaign in release data, Then there is not lists", async () => {
+    it("When there is not campaign in release data, Then there is not lists", () => {
         release_data = {
             id: 2,
             planning: {
@@ -93,12 +94,12 @@ describe("TestManagement", () => {
             campaign: null,
         } as MilestoneData;
 
-        const wrapper = await getPersonalWidgetInstance();
+        const wrapper = getPersonalWidgetInstance();
         expect(wrapper.find("[data-test=display-ttm]").exists()).toBe(false);
     });
 
-    it("When component is renderer, Then there is a div element with id of release", async () => {
-        const wrapper = await getPersonalWidgetInstance();
+    it("When component is renderer, Then there is a div element with id of release", () => {
+        const wrapper = getPersonalWidgetInstance();
         expect(wrapper.element).toMatchSnapshot();
     });
 });
