@@ -59,7 +59,7 @@ import type { MilestoneData } from "../../type";
 import { FetchWrapperError } from "@tuleap/tlp-fetch";
 import { is_testplan_activated } from "../../helpers/test-management-helper";
 import { useStore } from "../../stores/root";
-import { useGettext } from "@tuleap/vue2-gettext-composition-helper";
+import { useGettext } from "vue3-gettext";
 
 const root_store = useStore();
 
