@@ -17,25 +17,25 @@
  * along with Tuleap. If not, see <http://www.gnu.org/licenses/>.
  */
 
-import type { Wrapper } from "@vue/test-utils";
+import type { VueWrapper } from "@vue/test-utils";
 import { shallowMount } from "@vue/test-utils";
 import ReleaseHeader from "./ReleaseHeader.vue";
 import type { MilestoneData, Pane } from "../../../type";
 import { setUserLocale } from "../../../helpers/user-locale-helper";
-import { createReleaseWidgetLocalVue } from "../../../helpers/local-vue-for-test";
 import ReleaseHeaderRemainingDays from "./ReleaseHeaderRemainingDays.vue";
 import ReleaseHeaderRemainingPoints from "./ReleaseHeaderRemainingPoints.vue";
 import PastReleaseHeaderInitialPoints from "./PastReleaseHeaderInitialPoints.vue";
 import PastReleaseHeaderTestsDisplayer from "./PastReleaseHeaderTestsDisplayer.vue";
+import { getGlobalTestOptions } from "../../../helpers/global-options-for-test";
 
 describe("ReleaseHeader", () => {
-    async function getPersonalWidgetInstance(
+    function getPersonalWidgetInstance(
         start_date: null | string,
         is_loading: boolean,
         label: string,
         is_past_release: boolean,
         additional_panes: Array<Pane>,
-    ): Promise<Wrapper<Vue, Element>> {
+    ): VueWrapper<InstanceType<typeof ReleaseHeader>> {
         const release_data = {
             label,
             id: 2,
@@ -52,17 +52,19 @@ describe("ReleaseHeader", () => {
                 isLoading: is_loading,
                 isPastRelease: is_past_release,
             },
-            localVue: await createReleaseWidgetLocalVue(),
+            global: {
+                ...getGlobalTestOptions(),
+            },
         };
 
         return shallowMount(ReleaseHeader, component_options);
     }
 
     describe("Display arrow between dates", () => {
-        it("When there are a start date and end date, Then an arrow is displayed", async () => {
+        it("When there are a start date and end date, Then an arrow is displayed", () => {
             setUserLocale("en-US");
 
-            const wrapper = await getPersonalWidgetInstance(
+            const wrapper = getPersonalWidgetInstance(
                 new Date("2017-01-22T13:42:08+02:00").toDateString(),
                 false,
                 "mile",
@@ -73,38 +75,38 @@ describe("ReleaseHeader", () => {
             expect(wrapper.find("[data-test=display-arrow]").exists()).toBe(true);
         });
 
-        it("When there isn't a start date of a release, Then there isn't an arrow", async () => {
-            const wrapper = await getPersonalWidgetInstance(null, false, "mile", false, []);
+        it("When there isn't a start date of a release, Then there isn't an arrow", () => {
+            const wrapper = getPersonalWidgetInstance(null, false, "mile", false, []);
             expect(wrapper.find("[data-test=display-arrow]").exists()).toBe(false);
         });
     });
 
-    it("When the widget is loading, Then there is a skeleton instead of points", async () => {
-        const wrapper = await getPersonalWidgetInstance(null, true, "mile", false, []);
+    it("When the widget is loading, Then there is a skeleton instead of points", () => {
+        const wrapper = getPersonalWidgetInstance(null, true, "mile", false, []);
         expect(wrapper.find("[data-test=display-skeleton]").exists()).toBe(true);
     });
 
-    it("When release's title contains '>', Then '>' is displayed", async () => {
-        const wrapper = await getPersonalWidgetInstance(null, true, "1 > 2", false, []);
+    it("When release's title contains '>', Then '>' is displayed", () => {
+        const wrapper = getPersonalWidgetInstance(null, true, "1 > 2", false, []);
         expect(wrapper.get("[data-test=title-release]").text()).toBe("1 > 2");
     });
 
     describe("Display PastReleaseHeader", () => {
-        it("When the release is not past, Then ReleaseHeaderRemaining components are displayed", async () => {
-            const wrapper = await getPersonalWidgetInstance(null, false, "mile", false, []);
+        it("When the release is not past, Then ReleaseHeaderRemaining components are displayed", () => {
+            const wrapper = getPersonalWidgetInstance(null, false, "mile", false, []);
             expect(wrapper.findComponent(ReleaseHeaderRemainingDays).exists()).toBe(true);
             expect(wrapper.findComponent(ReleaseHeaderRemainingPoints).exists()).toBe(true);
             expect(wrapper.findComponent(PastReleaseHeaderInitialPoints).exists()).toBe(false);
         });
 
-        it("When the release is past, Then PastReleaseHeaderInitialPoints component are displayed", async () => {
-            const wrapper = await getPersonalWidgetInstance(null, false, "mile", true, []);
+        it("When the release is past, Then PastReleaseHeaderInitialPoints component are displayed", () => {
+            const wrapper = getPersonalWidgetInstance(null, false, "mile", true, []);
             expect(wrapper.findComponent(ReleaseHeaderRemainingDays).exists()).toBe(false);
             expect(wrapper.findComponent(ReleaseHeaderRemainingPoints).exists()).toBe(false);
             expect(wrapper.findComponent(PastReleaseHeaderInitialPoints).exists()).toBe(true);
         });
 
-        it("When the release is past and TestPlan is enabled, Then PastReleaseHeaderTestsDisplayer component are displayed", async () => {
+        it("When the release is past and TestPlan is enabled, Then PastReleaseHeaderTestsDisplayer component are displayed", () => {
             const additional_panes = [
                 {
                     icon_name: "fa-check",
@@ -114,13 +116,7 @@ describe("ReleaseHeader", () => {
                 },
             ];
 
-            const wrapper = await getPersonalWidgetInstance(
-                null,
-                false,
-                "mile",
-                true,
-                additional_panes,
-            );
+            const wrapper = getPersonalWidgetInstance(null, false, "mile", true, additional_panes);
             expect(wrapper.findComponent(ReleaseHeaderRemainingDays).exists()).toBe(false);
             expect(wrapper.findComponent(ReleaseHeaderRemainingPoints).exists()).toBe(false);
             expect(wrapper.findComponent(PastReleaseHeaderInitialPoints).exists()).toBe(true);
