@@ -55,7 +55,7 @@
 <script setup lang="ts">
 import { computed } from "vue";
 import type { MilestoneData } from "../../../type";
-import { useGettext } from "@tuleap/vue2-gettext-composition-helper";
+import { useGettext } from "vue3-gettext";
 
 const { $ngettext, $gettext } = useGettext();
 
